update code folder ManhLN

+++ WIP/Users/ManhLN/Source Code/DDL_CapstoneProject/DDL_CapstoneProject/Controllers/ApiControllers/ProjectApiController.cs
@@ -27,12 +27,12 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             try
@@ -43,17 +43,19 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = id });
         }
 
         // PUT: api/ProjectApi/EditProject  
-        [ResponseType(typeof(ProjectDetailDTO))]
+        [ResponseType(typeof(ProjectEditDTO))]
         [HttpPost]
         public IHttpActionResult EditProjectBasic()
         {
             var httpRequest = HttpContext.Current.Request;
+
+            ProjectEditDTO updateProject = null;
 
             if (httpRequest.Form.Count <= 0)
             {
@@ -61,13 +63,26 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
 
             var projectJson = httpRequest.Form["project"];
-            var project = new JavaScriptSerializer().Deserialize<ProjectDetailDTO>(projectJson);
+            var project = new JavaScriptSerializer().Deserialize<ProjectEditDTO>(projectJson);
+
+            if (project == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             string imageName = "img_" + project.ProjectCode;
             var file = httpRequest.Files["file"];
             var uploadImageName = CommonUtils.UploadImage(file, imageName, DDLConstants.FileType.PROJECT);
 
-            var updateProject = ProjectRepository.Instance.EditProjectBasic(project, uploadImageName);
+            try
+            {
+                updateProject = ProjectRepository.Instance.EditProjectBasic(project, uploadImageName);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -77,22 +92,67 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpPut]
         public IHttpActionResult EditProjectStory(ProjectStoryDTO project)
         {
+            ProjectStoryDTO updateProject = null;
+
             if (project == null)
             {
                 return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            var updateProject = ProjectRepository.Instance.EditProjectStory(project);
+            try
+            {
+                updateProject = ProjectRepository.Instance.EditProjectStory(project);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
 
         // GET: api/ProjectApi/GetProject/:id
         [HttpGet]
-        [ResponseType(typeof(ProjectDetailDTO))]
+        [ResponseType(typeof(ProjectEditDTO))]
         public IHttpActionResult GetProject(int id)
         {
-            var project = ProjectRepository.Instance.GetProjectBasic(id);
+            ProjectEditDTO project = null;
+
+            var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            if (!ModelState.IsValid)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            if (User.Identity == null || !User.Identity.IsAuthenticated)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            try
+            {
+                project = ProjectRepository.Instance.GetProjectBasic(id, currentUser.DDL_UserID);
+                if (project.ImageUrl != string.Empty)
+                {
+                    project.ImageUrl = DDLConstants.FileType.PROJECT + project.ImageUrl;
+                }
+            }
+            catch (ProjectNotFoundException)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Không tìm thấy dự án!", Type = DDLConstants.HttpMessageType.NOT_FOUND });
+            }
+            catch (Exception)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = project });
         }
@@ -102,7 +162,16 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(ProjectStoryDTO))]
         public IHttpActionResult GetProjectStory(int id)
         {
-            var project = ProjectRepository.Instance.GetProjectStory(id);
+            ProjectStoryDTO project = null;
+            try
+            {
+                project = ProjectRepository.Instance.GetProjectStory(id);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = project });
         }
@@ -112,7 +181,17 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(RewardPkgDTO))]
         public IHttpActionResult GetRewardPkg(int id)
         {
-            var rewardPkg = RewardPkgRepository.Instance.GetRewardPkg(id);
+            List<RewardPkgDTO> rewardPkg;
+
+            try
+            {
+                rewardPkg = RewardPkgRepository.Instance.GetRewardPkg(id);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = rewardPkg });
         }
@@ -127,12 +206,12 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             try
@@ -141,7 +220,7 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = rewardPkg });
         }
@@ -151,12 +230,21 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpPut]
         public IHttpActionResult EditRewardPkg(List<RewardPkgDTO> rewardPkg)
         {
+            bool result = false;
             if (rewardPkg == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            var result = RewardPkgRepository.Instance.EditRewardPkg(rewardPkg);
+            try
+            {
+                result = RewardPkgRepository.Instance.EditRewardPkg(rewardPkg);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -166,6 +254,13 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpDelete]
         public IHttpActionResult DeleteRewardPkg(int id)
         {
+            var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
             try
             {
                 var result = RewardPkgRepository.Instance.DeleteRewardPkg(id);
@@ -173,7 +268,7 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
@@ -184,7 +279,23 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(UpdateLogDTO))]
         public IHttpActionResult GetUpdateLog(int id)
         {
-            var updateLog = UpdateLogRepository.Instance.GetUpdateLog(id);
+            List<UpdateLogDTO> updateLog = null;
+            var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            try
+            {
+                updateLog = UpdateLogRepository.Instance.GetUpdateLog(id);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = updateLog });
         }
@@ -195,25 +306,26 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         public IHttpActionResult CreateUpdateLog(int id, UpdateLogDTO newUpdateLog)
         {
             var currentUser = getCurrentUser();
+            var updateLog = new UpdateLog();
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            //try
-            //{
-            var updateLog = UpdateLogRepository.Instance.CreateUpdateLog(id, newUpdateLog);
-            //}
-            //catch (Exception)
-            //{
-            //    return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
-            //}
+            try
+            {
+                updateLog = UpdateLogRepository.Instance.CreateUpdateLog(id, newUpdateLog);
+            }
+            catch (Exception)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = updateLog });
         }
 
@@ -222,12 +334,21 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpPut]
         public IHttpActionResult EditUpdateLog(List<UpdateLogDTO> updateLog)
         {
+
             if (updateLog == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            var result = UpdateLogRepository.Instance.EditUpdateLog(updateLog);
+            try
+            {
+                var result = UpdateLogRepository.Instance.EditUpdateLog(updateLog);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -248,24 +369,8 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
-            }
-
-            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
-        }
-
-        // Post: Upload Image
-        [HttpPost]
-        public IHttpActionResult fileUpload(string ProjectCode)
-        {
-            var httpRequest = HttpContext.Current.Request;
-            if (httpRequest.Files.Count <= 0)
-            {
                 return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
-            string imageName = "img_" + ProjectCode;
-            var file = httpRequest.Files[0];
-            var uploadImageName = CommonUtils.UploadImage(file, imageName, DDLConstants.FileType.PROJECT);
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -292,6 +397,31 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         {
             var listStatisticForHome = ProjectRepository.Instance.GetStatisticListForHome();
             return Ok(new HttpMessageDTO { Status = "success", Data = listStatisticForHome });
+        }
+
+        // PUT: api/ProjectApi/EditProjectStory  
+        [ResponseType(typeof(ProjectEditDTO))]
+        [HttpPut]
+        public IHttpActionResult SubmitProject(ProjectEditDTO project)
+        {
+            ProjectEditDTO updateProject = null;
+
+            if (project == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            try
+            {
+                updateProject = ProjectRepository.Instance.SubmitProject(project);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
 
         //Trungvn
@@ -339,6 +469,27 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             });
         }
 
+        //  17/10/2015 - MaiCTP - Get BAckedProject
+        public IHttpActionResult GetBackedProject()
+        {
+            var listBacked = ProjectRepository.Instance.GetBackedProject(User.Identity.Name);
+            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = listBacked });
+        }
+
+
+        //  18/10/2015 - MaiCTP - Get StarredProject
+        public IHttpActionResult GetStarredProject()
+        {
+            var listStarred = ProjectRepository.Instance.GetStarredProject(User.Identity.Name);
+            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = listStarred });
+        }
+
+        //  18/10/2015 - MaiCTP - Get CreatedProject
+        public IHttpActionResult GetCreatedProject()
+        {
+            var listCreated = ProjectRepository.Instance.GetCreatedProject(User.Identity.Name);
+            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = listCreated });
+        }
         #region Comment Functions
 
         /// <summary>
